+++ api/src/main/scala/pl/touk/nussknacker/engine/api/component/ComponentProvider.scala
@@ -64,6 +64,8 @@ trait ComponentProvider {
 
   def isCompatible(version: NussknackerVersion): Boolean
 
+  def isAutoLoaded: Boolean = false
+
 }
 
 object NussknackerVersion {