+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/component/ComponentExtractor.scala
@@ -22,9 +22,14 @@ case class ComponentExtractor(classLoader: ClassLoader, nussknackerVersion: Nuss
 
   private lazy val providers = ScalaServiceLoader.load[ComponentProvider](classLoader).map(p => p.providerName -> p).toMap
 
-  private def loadCorrectComponents(config: Config): Map[String, (ComponentProviderConfig, ComponentProvider)] = {
-
+  private def loadCorrectProviders(config: Config): Map[String, (ComponentProviderConfig, ComponentProvider)] = {
     val componentsConfig = config.getAs[Map[String, ComponentProviderConfig]](componentConfigPath).getOrElse(Map.empty)
+    val manuallyLoadedProvidersWithConfig = loadManuallyLoadedProviders(componentsConfig)
+    val autoLoadedProvidersWithConfig = loadAutoLoadedProviders(componentsConfig, manuallyLoadedProvidersWithConfig)
+    manuallyLoadedProvidersWithConfig ++ autoLoadedProvidersWithConfig
+  }
+
+  private def loadManuallyLoadedProviders(componentsConfig: Map[String, ComponentProviderConfig]) = {
     componentsConfig.filterNot(_._2.disabled).map {
       case (name, providerConfig: ComponentProviderConfig) =>
         val providerName = providerConfig.providerType.getOrElse(name)
@@ -36,13 +41,26 @@ case class ComponentExtractor(classLoader: ClassLoader, nussknackerVersion: Nuss
     }
   }
 
+  private def loadAutoLoadedProviders(componentsConfig: Map[String, ComponentProviderConfig], manuallyLoadedProvidersWithConfig: Map[String, (ComponentProviderConfig, ComponentProvider)]) = {
+    val manuallyLoadedProviders = manuallyLoadedProvidersWithConfig.values.map(_._2).toSet
+    val autoLoadedProvidersWithConfig = providers.values
+      .filter(provider => provider.isAutoLoaded && !manuallyLoadedProviders.contains(provider) && !componentsConfig.get(provider.providerName).exists(_.disabled))
+      .map { provider =>
+        if (!provider.isCompatible(nussknackerVersion)) {
+          throw new IllegalArgumentException(s"Auto-loaded component provider ${provider.providerName} is not compatible with $nussknackerVersion, please use correct component provider version or disable it explicitly.")
+        }
+        provider.providerName -> (ComponentProviderConfig(providerType = None, componentPrefix = None), provider)
+      }
+    autoLoadedProvidersWithConfig
+  }
+
   def extract(processObjectDependencies: ProcessObjectDependencies): Map[String, WithCategories[Component]] =
-    loadCorrectComponents(processObjectDependencies.config).map {
+    loadCorrectProviders(processObjectDependencies.config).map {
       case (_, (config, provider)) => extractOneProviderConfig(config, provider, processObjectDependencies)
     }.reduceUnique
 
   def loadAdditionalConfig(inputConfig: Config, configWithDefaults: Config): Config = {
-    val resolvedConfigs = loadCorrectComponents(configWithDefaults).map {
+    val resolvedConfigs = loadCorrectProviders(configWithDefaults).map {
       case (name, (config, provider)) => name -> provider.resolveConfigForExecution(config.config)
     }
     resolvedConfigs.foldLeft(inputConfig) {