+++ interpreter/src/test/scala/pl/touk/nussknacker/engine/component/ComponentExtractorTest.scala
@@ -27,7 +27,10 @@ class ComponentExtractorTest extends FunSuite with Matchers {
   private val loader = new DefaultModelConfigLoader
 
   test("should discover services") {
-    val components = extractComponents[Service]("components.dynamicTest.valueCount" -> 7)
+    val components = extractComponents[Service]("components" -> Map(
+      "dynamicTest" -> Map("valueCount" -> 7),
+      "auto" -> Map("disabled" -> true)
+    ))
     components shouldBe (1 to 7).map(i => s"component-v$i" -> WithCategories(DynamicService(s"v$i"), None, SingleComponentConfig.zero)).toMap
   }
 
@@ -35,7 +38,8 @@ class ComponentExtractorTest extends FunSuite with Matchers {
     val components = extractComponents[Service](
       "components" -> Map(
         "dynamic1" -> Map("providerType" -> "dynamicTest", "valueCount" -> 2),
-        "dynamic2" -> Map("providerType" -> "dynamicTest", "componentPrefix" -> "t1-", "valueCount" -> 3)
+        "dynamic2" -> Map("providerType" -> "dynamicTest", "componentPrefix" -> "t1-", "valueCount" -> 3),
+        "auto" -> Map("disabled" -> true)
     ))
 
     components shouldBe ((1 to 2).map(i => s"component-v$i" -> WithCategories(DynamicService(s"v$i"), None, SingleComponentConfig.zero)) ++
@@ -56,7 +60,8 @@ class ComponentExtractorTest extends FunSuite with Matchers {
     val components = extractComponents[Service](
       "components" -> Map(
         "dynamic1" -> Map("providerType" -> "dynamicTest", "disabled" -> true),
-        "dynamic2" -> Map("providerType" -> "dynamicTest", "componentPrefix" -> "t1-", "valueCount" -> 1)
+        "dynamic2" -> Map("providerType" -> "dynamicTest", "componentPrefix" -> "t1-", "valueCount" -> 1),
+        "auto" -> Map("disabled" -> true)
     ))
     components shouldBe Map("t1-component-v1" -> WithCategories(DynamicService("v1"), None, SingleComponentConfig.zero))
   }
@@ -70,11 +75,26 @@ class ComponentExtractorTest extends FunSuite with Matchers {
     }.getMessage should include(s"is not compatible with NussknackerVersion(${largeVersionNumber.toString})")
   }
 
+  test("should load auto loadable component") {
+    val components = extractComponents[Service]()
+    components shouldBe Map("auto-component" -> WithCategories(AutoService, None, SingleComponentConfig.zero))
+  }
+
+  test("should skip incompatible auto loadable providers") {
+    //see DynamicProvider.isCompatible
+    val largeVersionNumber = new Semver(s"$largeMajorVersion.2.3")
+    intercept[IllegalArgumentException] {
+      extractComponents[Service](Map.empty[String, Any], (cl:ClassLoader) => ComponentExtractor(cl, NussknackerVersion(largeVersionNumber)))
+    }.getMessage should include(s"is not compatible with NussknackerVersion(${largeVersionNumber.toString})")
+  }
+
   private def extractComponents[T <: Component](map: (String, Any)*): Map[String, WithCategories[T]] =
     extractComponents(map.toMap, ComponentExtractor(_))
 
   private def extractComponents[T <: Component](map: Map[String, Any], makeExtractor: ClassLoader => ComponentExtractor): Map[String, WithCategories[T]] = {
-    ClassLoaderWithServices.withCustomServices(List((classOf[ComponentProvider], classOf[DynamicProvider])), getClass.getClassLoader) { cl =>
+    ClassLoaderWithServices.withCustomServices(List(
+      (classOf[ComponentProvider], classOf[DynamicProvider]),
+      (classOf[ComponentProvider], classOf[AutoLoadedProvider])), getClass.getClassLoader) { cl =>
       val extractor = makeExtractor(cl)
       val resolved = loader.resolveInputConfigDuringExecution(fromMap(map.toSeq: _*), cl)
       extractor.extract(ProcessObjectDependencies(resolved.config, DefaultNamespacedObjectNaming)).mapValues(k => k.copy(value = k.value.asInstanceOf[T]))
@@ -117,3 +137,23 @@ case class DynamicService(valueToReturn: String) extends Service {
   }
 }
 
+class AutoLoadedProvider extends ComponentProvider {
+  override def providerName: String = "auto"
+
+  override def resolveConfigForExecution(config: Config): Config = config
+
+  override def create(config: Config, dependencies: ProcessObjectDependencies): List[ComponentDefinition] = List(
+    ComponentDefinition("auto-component", AutoService)
+  )
+
+  override def isCompatible(version: NussknackerVersion): Boolean = version.value.getMajor < ComponentExtractorTest.largeMajorVersion
+
+  override def isAutoLoaded: Boolean = true
+}
+
+object AutoService extends Service {
+  @MethodToInvoke
+  def invoke(): Future[String] = {
+    ???
+  }
+}