+++ docs/Changelog.md
@@ -35,6 +35,7 @@ a custom DB metadata provider that extends the standard JDBC Driver with missing
 * [#2331](https://github.com/TouK/nussknacker/pull/2331) Refactor `nussknacker-avro-flink-util` module. Move avro specific classes to new `nussknacker-avro-util` module. 
 * [#2344](https://github.com/TouK/nussknacker/pull/2344) Redesign of `#DATE` and `#DATE_FORMAT` utilities.
 * [#2305](https://github.com/TouK/nussknacker/pull/2305) Enhancement: change `processingTypeToDashboard` configuration to `scenarioTypeToDashboard`
+* [#2374](https://github.com/TouK/nussknacker/pull/2374) Auto-loaded `CompoentProvider`s
 
 1.0.0
 ------------------------